Added a new usestate error and error object
AddUser.js: Added a new usesate error to store different values of error
objects that is created according to the user validation. Depending on
the error detected one error object is cretaed in the "if"
statements. In addition, if the error object is not defined this
prevents ErrorModal from showing.